pyproject.toml
[project]
name = "python-in-data-project"
version = "0.1.0"
description = ""
authors = [
    {name = "NadymB",email = "nguyenleduydang2010@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker (>=40.1.2,<41.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandas (>=3.0.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "python_in_data_project", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
